% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/byID.interactions.R
\name{byID.interactions}
\alias{byID.interactions}
\title{byID.interactions {WayTOols}}
\usage{
byID.interactions(logfile_data, ID = "", ...)
}
\arguments{
\item{logfile_data}{the data from the object's data frame. WayTOols is built
specifically to deal with logfile_data generated from the WayTO mobile application.}

\item{ID}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{...}{Arguments to be passed to methods, such as na.rm=TRUE}
}
\description{
Take a sequence of logfile data frame arguments and combine
by "ID", "Zoomed", "Panned" and "WithOffScreenMarkers" by columns.
}
\details{
This function cleans up the original data frame "logfile_data" to include
only the events corresponding to user interaction with the digital map. This function
enables the choice of which ID should be displayed in the condensed data frame.
}
\examples{
## when na.rm = TRUE
data(logfile_data)
byID.interactions(logfile_data, ID="13w")
}
\seealso{
byTime.entries, mean.interactions, plot.interactions, sum.interactions
}

