% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/byTime.entries.R
\name{byTime.entries}
\alias{byTime.entries}
\title{byTime.entries {WayTOols}}
\usage{
byTime.entries(logfile_data, ID = "", ...)
}
\arguments{
\item{logfile_data}{the data from the object's data frame. WayTOols is built
specifically to deal with logfile_data generated from the WayTO mobile application.}

\item{ID}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{...}{Arguments to be passed to methods, such as na.rm=TRUE}
}
\description{
Counts the number of entries by a factor of time in the logfile.
}
\details{
This function provides users with the opportunity to count the number of
interactions by the user ID. This is important because the Time variable is the only
constant variable over the entire dataset.
}
\note{
test
}
\examples{
## count the number of entries by time in ID=4a
data(logfile_data)
byTime.entries(logfile_data, ID="4a")
}
\seealso{
byID.interactions, mean.interactions, plot.interactions, sum.interactions
}

