% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_logfile.R
\docType{data}
\name{logfile_data}
\alias{logfile_data}
\title{logfile_data}
\format{The logfile_data data frame has 3137 rows and 10 columns on map interactions at various points in time.}
\usage{
data(logfile_data)
}
\description{
logfile_data is a dataset that was saved as a .csv file after reading .txt files, otherwise known as logfiles, generated by the mobile application "WayTO" into R. This dataset was obtained through a two-month experiment undertaken by the Spatial Intelligence Lab at the University of Münster.
}
\details{
Description

ID Unique ID given to each participant that makes up the file name for each logfile. Each ID contains a number and a letter. The number corresponds with the order of participants, and the letter corresponds to
1. The researcher (a/b - Researcher 1; w/wo - Researcher 2) and
2. The viewability of off-screen markers (a/w = with on-screen markers, b/wo = without off-screen markers)
}
\examples{
## load data
# data(logfile_data)
x <- as.numeric(logfile_data$Zoom, ID="3w")
x <- head(x, 50)
mean(x)
}
\keyword{datasets}

