% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean.interactions.R
\name{mean.interactions}
\alias{mean}
\alias{mean.interactions}
\title{mean.interactions {WayTOols}}
\usage{
mean.interactions(logfile_data, ID = "", ...)
}
\arguments{
\item{logfile_data}{the data from the object's data frame. WayTOols is built
specifically to deal with logfile_data generated from the WayTO mobile application.}

\item{ID}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{...}{Arguments to be passed to methods, such as na.rm=TRUE}
}
\description{
Returns the average amount of Zoom and Pan interactions by user
}
\details{
For a given ID, the average zoom level and average percentage of panning for the digital map are returned. This information is useful in that we can see the degree of panning (is it a lot? is the participant interacting a lot with the map?) and zooming (what is the average zoom level? Are they mostly zoomed in close to map elements, or far away?)
}
\examples{
data(logfile_data)
mean.interactions(logfile_data, ID="6wo")
}
\seealso{
byTime.entries, byID.interactions, plot.interactions, sum.interactions
}

