% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.interactions.R
\name{plot.interactions}
\alias{plot}
\alias{plot.interactions}
\title{plot.interactions {WayTOols}}
\usage{
plot.interactions(logfile_data, ID = "", ...)
}
\arguments{
\item{logfile_data}{the data from the object's data frame. WayTOols is built
specifically to deal with logfile_data generated from the WayTO mobile application.}

\item{ID}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{...}{Arguments to be passed to methods, such as na.rm=TRUE}
}
\description{
Plots the locations where a specific user zoomed in or out of the digital map, and visualizes the zoom by size of zoom.
}
\details{
Plots are generated to show the location-based zoom level interactions by user. This is interesting when analyzing the logfile data to see if users are constantly zooming throughout their experiment navigation
}
\note{
requires ggplot2 and ggmap
}
\examples{
## Plot interactions on a map and show boxplot
library(ggplot2)
library(ggmap)

plot.interactions(logfile_data, ID="4a")

ggplot(logfile_data, aes(factor(ID), Zoom), na.rm=TRUE) +
geom_boxplot() +
xlab("User ID")+
ggtitle("Zoom factor across different Users")
}
\seealso{
byTime.entries, byID.interactions, mean.interactions, sum.interactions
}

