% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sum.interactions.R
\name{sum.interactions}
\alias{sum}
\alias{sum.interactions}
\title{sum.interactions {WayTOols}}
\usage{
sum.interactions(logfile_data, ID = "", ...)
}
\arguments{
\item{logfile_data}{the data from the object's data frame. WayTOols is built
specifically to deal with logfile_data generated from the WayTO mobile application.}

\item{ID}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{...}{Arguments to be passed to methods, such as na.rm=TRUE}
}
\description{
Provides the sum of interactions per user by Zoom and Pan factor
}
\details{
When the ID and logfile data are provided, the total amount of Zoom and Pan interactions are returned by this function.
}
\examples{
require(dplyr)
data(logfile_data)
sum.interactions(logfile_data, ID="6wo")
}
\seealso{
byTime.entries, byID.interactions, plot.interactions, mean.interactions
}

