% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countZoom.R, R/meanZoom.R
\docType{class}
\name{User-class}
\alias{User-class}
\alias{countZoom}
\alias{meanZoom}
\title{countZoom {WayTOols}}
\usage{
countZoom(object)

meanZoom(object)
}
\description{
Counts number of Zooms per user

Averages number of Zooms per user
}
\details{
The sum of the number of zoom interactions per user.

The meanZoom method and the countZoom method are able to calculate the average and total zoom for a specific participant. The participant's data must first be loaded into the workspace.
}
\section{Slots}{

\describe{
\item{\code{object}}{A hard-coded argument for User ID and its average ID. See example for details}

\item{\code{User}}{class specifies the ID and interaction parameters}

\item{\code{data}}{the logfile data read as a data frame.}

\item{\code{ID}}{Each logfile has a unique ID corresponding to a participant in the experiment}

\item{\code{Zoom}}{is the value of zoom level (e.g. 14.0) measured at each map interaction}

\item{\code{Zoomed}}{returns a TRUE/FALSE value for each interaction. It enables the user to know whether the interaction was a Zoom into the map or not.}
}}
\note{
see utils

A User.ID must first be hard coded before the meanZoom or countZoom methods can return the average zoom level.
}
\section{S4 Method}{

}
\examples{
## Set User ID
#User.13w <- new("User",
#ID = "ID",
#Zoom = logfile_data$Zoom[logfile_data$ID == "13w"] ,
#Zoomed = logfile_data$Zoomed[logfile_data$ID == "13w"] )

## countZoom(User.13w)
## Set User ID
#User.9w <- new("User",
#ID = "ID",
#Zoom = logfile_data$Zoom[logfile_data$ID == "9w"] ,
#Zoomed = logfile_data$Zoomed[logfile_data$ID == "9w"] )

## meanZoom(User.9w)
}
\seealso{
utils

countZoom
}

